requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
shap>=0.42.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.0.0
anthropic>=0.7.0
dice-ml>=0.10.0
PyYAML>=6.0
requests>=2.31.0
textblob>=0.17.1
nltk>=3.8.0

